% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitCTVARMx-methods-fit-ct-var-id-mx.R
\name{coef.fitctvaridmx}
\alias{coef.fitctvaridmx}
\title{Parameter Estimates}
\usage{
\method{coef}{fitctvaridmx}(object, iota = FALSE, sigma = FALSE, theta = FALSE, ...)
}
\arguments{
\item{object}{Object of class \code{fitctvaridmx}.}

\item{iota}{Logical.
If \code{iota = TRUE},
include estimates of the \code{iota} vector, if available.
If \code{iota = FALSE},
exclude estimates of the \code{iota} vector.}

\item{sigma}{Logical.
If \code{sigma = TRUE},
include estimates of the \code{sigma} matrix, if available.
If \code{sigma = FALSE},
exclude estimates of the \code{sigma} matrix.}

\item{theta}{Logical.
If \code{theta = TRUE},
include estimates of the \code{theta} matrix, if available.
If \code{theta = FALSE},
exclude estimates of the \code{theta} matrix.}

\item{...}{additional arguments.}
}
\value{
Returns a list of vectors of parameter estimates.
}
\description{
Parameter Estimates
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
