% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitCTVARMx-package.R
\docType{package}
\name{fitCTVARMx-package}
\alias{fitCTVARMx}
\alias{fitCTVARMx-package}
\title{fitCTVARMx: Fit The Continuous-Time Vector Autoregressive Model}
\description{
Fit the continuous-time vector autoregressive model using the 'OpenMx' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jeksterslab/fitCTVARMx}
  \item \url{https://jeksterslab.github.io/fitCTVARMx/}
  \item Report bugs at \url{https://github.com/jeksterslab/fitCTVARMx/issues}
}

}
\author{
\strong{Maintainer}: Ivan Jacob Agaloos Pesigan \email{r.jeksterslab@gmail.com} (\href{https://orcid.org/0000-0003-4818-8420}{ORCID}) [copyright holder]

}
\keyword{internal}
