% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitCTVARMx-methods-fit-ct-var-id-mx.R
\name{print.fitctvaridmx}
\alias{print.fitctvaridmx}
\title{Print Method for Object of Class \code{fitctvaridmx}}
\usage{
\method{print}{fitctvaridmx}(x, means = TRUE, ...)
}
\arguments{
\item{x}{an object of class \code{fitctvaridmx}.}

\item{means}{Logical.
If \code{means = TRUE}, return means.
Otherwise, the function returns raw estimates.}

\item{...}{further arguments.}
}
\description{
Print Method for Object of Class \code{fitctvaridmx}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
