% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitCTVARMx-methods-fit-ct-var-mx.R
\name{print.fitctvarmx}
\alias{print.fitctvarmx}
\title{Print Method for Object of Class \code{fitctvarmx}}
\usage{
\method{print}{fitctvarmx}(x, ...)
}
\arguments{
\item{x}{an object of class \code{fitctvarmx}.}

\item{...}{further arguments.}
}
\description{
Print Method for Object of Class \code{fitctvarmx}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
