% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitCTVARMx-methods-fit-ct-var-id-mx.R
\name{summary.fitctvaridmx}
\alias{summary.fitctvaridmx}
\title{Summary Method for Object of Class \code{fitctvaridmx}}
\usage{
\method{summary}{fitctvaridmx}(object, means = TRUE, ...)
}
\arguments{
\item{object}{an object of class \code{fitctvaridmx}.}

\item{means}{Logical.
If \code{means = TRUE}, return means.
Otherwise, the function returns raw estimates.}

\item{...}{further arguments.}
}
\description{
Summary Method for Object of Class \code{fitctvaridmx}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
