% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitCTVARMx-methods-fit-ct-var-mx.R
\name{summary.fitctvarmx}
\alias{summary.fitctvarmx}
\title{Summary Method for Object of Class \code{fitctvarmx}}
\usage{
\method{summary}{fitctvarmx}(object, ...)
}
\arguments{
\item{object}{an object of class \code{fitctvarmx}.}

\item{...}{further arguments.}
}
\description{
Summary Method for Object of Class \code{fitctvarmx}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
