% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitCTVARMx-methods-fit-ct-var-id-mx.R
\name{vcov.fitctvaridmx}
\alias{vcov.fitctvaridmx}
\title{Sampling Covariance Matrix of the Parameter Estimates}
\usage{
\method{vcov}{fitctvaridmx}(object, sigma = FALSE, theta = FALSE, ...)
}
\arguments{
\item{object}{Object of class \code{fitctvaridmx}.}

\item{sigma}{Logical.
If \code{sigma = TRUE},
include estimates of the \code{sigma} matrix, if available.
If \code{sigma = FALSE},
exclude estimates of the \code{sigma} matrix.}

\item{theta}{Logical.
If \code{theta = TRUE},
include estimates of the \code{theta} matrix, if available.
If \code{theta = FALSE},
exclude estimates of the \code{theta} matrix.}

\item{...}{additional arguments.}
}
\value{
Returns a list of sampling variance-covariance matrices.
}
\description{
Sampling Covariance Matrix of the Parameter Estimates
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
